ImP Setup - Reduced size of elapsed time intiger - Fixed some issues with time handling in adding to data

+++ src/timing/timer.h
@@ -20,8 +20,8 @@ public:
 		beg_ = clock_::now();
 	}
 
-	int64_t time() const {
-		return (int64_t) std::chrono::duration_cast<millisecs_>(clock_::now().time_since_epoch()).count();
+	int32_t time() const {
+		return (int32_t) std::chrono::duration_cast<millisecs_>(clock_::now().time_since_epoch()).count();
 	}
 
 	/**
@@ -38,8 +38,8 @@ public:
 		return ss.str();
 	}
 
-	int64_t elapsed() const {
-		return (int64_t) std::chrono::duration_cast<millisecs_>(clock_::now() - beg_).count();
+	int32_t elapsed() const {
+		return (int32_t) std::chrono::duration_cast<millisecs_>(clock_::now() - beg_).count();
 	}
 
 	static void sleep_ms(int ms) {
@@ -48,7 +48,7 @@ public:
 
 private:
 	typedef std::chrono::system_clock clock_;
-	typedef std::chrono::duration<uint64_t, std::milli> millisecs_;
+	typedef std::chrono::duration<uint32_t, std::milli> millisecs_;
 	std::chrono::time_point<clock_> beg_;
 };
